GroundProjectedEnv separated from Environment
- scalar prop renamed to scale. It's now Scale type from threlte instead of number
- docs and examples updated to reflect this change
