+++ apps/docs/src/examples/environment/Wrapper.svelte
@@ -10,7 +10,7 @@
 		files="shanghai_riverside_1k.hdr"
 		isBackground={true}
 		format="hdr"
-		groundProjection={{ radius: 200, height: 5, scalar: 100 }}
+		groundProjection={{ radius: 50, height: 5, scale: 50 }}
 	/>
 	<Scene />
 </Canvas>
